service/api/app/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile components/iam_session_broker/runtime/requirements.in
#
aws-lambda-powertools==3.4.0
    # via -r components/iam_session_broker/runtime/requirements.in
boto3==1.35.97
    # via -r components/iam_session_broker/runtime/requirements.in
boto3-stubs[dynamodb,sts]==1.35.97
    # via -r components/iam_session_broker/runtime/requirements.in
botocore==1.35.97
    # via
    #   boto3
    #   s3transfer
botocore-stubs==1.35.97
    # via boto3-stubs
cffi==1.17.1
    # via cryptography
cryptography==44.0.0
    # via pyjwt
jmespath==1.0.1
    # via
    #   aws-lambda-powertools
    #   boto3
    #   botocore
mypy-boto3-dynamodb==1.35.94
    # via boto3-stubs
mypy-boto3-sts==1.35.97
    # via boto3-stubs
pycparser==2.22
    # via cffi
pyjwt[crypto]==2.10.1
    # via -r components/iam_session_broker/runtime/requirements.in
python-dateutil==2.9.0.post0
    # via botocore
s3transfer==0.10.4
    # via boto3
six==1.17.0
    # via python-dateutil
types-awscrt==0.23.6
    # via botocore-stubs
types-s3transfer==0.10.4
    # via boto3-stubs
typing-extensions==4.12.2
    # via
    #   aws-lambda-powertools
    #   boto3-stubs
    #   mypy-boto3-dynamodb
    #   mypy-boto3-sts
urllib3==1.26.20
    # via botocore
